// frontend/src/services/Interfaces.ts
export  interface IPPP {
    id: string;
    medic_name: string;
    func_name: string;
    exam_name: string;
    agent_name: string;
    image: string;
    description: string;

}

export  interface IFuncionario {
    id: string;
    name: string;
}

export  interface IExames {
    id: string;
    name: string;
}

export  interface IAgentRisco {
    id: string;
    name: string;
}